fix(moe): use fp32 correction_bias in trtllm_fp8 kernel to match routing_logits
GLM 5.2 FP8 path cast correction_bias to bf16 (x.dtype) while
routing_logits was already fp32, causing a dtype mismatch inside
the flashinfer fused routing kernel. This led to precision loss in
softmax+topk expert routing. Align bias dtype to torch.float32,
matching the nvfp4 path's already-correct behavior.

Mirrors sglang PR #29783 for the same class of bug.

+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/moe/flashinfer/trtllm_fp8.py
@@ -139,7 +139,7 @@ if platform.is_nvidia:
         num_experts = getattr(w, "num_experts", local_experts)
         correction_bias = _routing_value(w, "correction_bias", None)
         routing_bias = (
-            correction_bias.to(x.dtype)
+            correction_bias.to(torch.float32)
             if isinstance(correction_bias, torch.Tensor)
             else None
         )
